requirements.txt
#Ask Ini to install python 3.11.x

easyocr==1.7.2
filelock==3.19.1
fsspec==2025.9.0
ImageIO==2.37.2
Jinja2==3.1.3
lazy_loader==0.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.5
ninja==1.13.0
numpy==1.26.4
opencv-python-headless==4.8.0.76
packaging==25.0
pillow==11.3.0
pyclipper==1.4.0
python-bidi==0.6.7
pytz==2025.2
PyYAML==6.0.3
scikit-image==0.25.2
scipy==1.16.3
shapely==2.1.2
sympy==1.14.0
tifffile==2025.10.16
torch==2.2.2
torchvision==0.17.2
typing_extensions==4.12.2